requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit>=1.39.0
requests==2.31.0
beautifulsoup4==4.12.2
openai>=1.6.1,<2.0.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
tiktoken>=0.5.2,<0.6.0
lxml>=5.3.0,<6.0
protobuf>=5.0,<7.0

